seperate classes into two different files
